(FLOW): Factored out common test-base for the max-flow algorithms;              Extracted perf-tests
